MAINT: windows changes for porting to IP7s

+++ src/SOCKITopenConnection.cpp
@@ -198,10 +198,10 @@ ExecuteSOCKITopenconnection(SOCKITopenconnectionRuntimeParamsPtr p)
 	}
 #endif
 
-	rc = connect(sockNum, servinfo->ai_addr, servinfo->ai_addrlen);
+	rc = connect(sockNum, servinfo->ai_addr, (int)servinfo->ai_addrlen);
 	//rc returns -1, with EINPROGRESS, because we set the socket to blocking.  Therefore you 
 	//have to test if the socket is writable with select
-	res = select(sockNum + 1, 0, &tempset,0, &timeout);
+	res = select((int) sockNum + 1, 0, &tempset,0, &timeout);
 	
 	if(res > 0 && FD_ISSET(sockNum, &tempset)){
 		if(!p->QFlagEncountered){
@@ -285,7 +285,7 @@ done:
 	FD_ZERO(&tempset);
 	if(!err && sockNum>0 && !err2){
 		err = SetOperationNumVar("V_flag", 0);
-		err = StoreNumericDataUsingVarName(p->IDVarName, sockNum,0);
+		err = StoreNumericDataUsingVarName(p->IDVarName, (double) sockNum, 0);
 	} else {
 		err = SetOperationNumVar("V_flag",1);
 		err = StoreNumericDataUsingVarName(p->IDVarName, -1, 0);
@@ -388,10 +388,10 @@ SOCKITopenconnectionF(SOCKITopenconnectionFStructPtr p)
 		goto done;
 #endif
 	
-	rc = connect(sockNum, servinfo->ai_addr, servinfo->ai_addrlen);
+	rc = connect(sockNum, servinfo->ai_addr, (int) servinfo->ai_addrlen);
 	//rc returns -1, with EINPROGRESS, because we set the socket to blocking.  Therefore you 
 	//have to test if the socket is writable with select
-	res = select(sockNum + 1, 0, &tempset,0, &timeout);
+	res = select((int) sockNum + 1, 0, &tempset,0, &timeout);
 	
 	if(res > 0 && FD_ISSET(sockNum, &tempset)){
 	} else {
@@ -454,7 +454,7 @@ done:
 	
 	FD_ZERO(&tempset);
 	if(!err && sockNum > 0 && !err2){
-		p->retval = sockNum;
+		p->retval = (double) sockNum;
 	} else {
 		p->retval = -1;
 	}
